doc_pipeline/pyproject.toml
[project]
name = "docs-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mwparserfromhell>=0.7.2",
    "mwxml>=0.3.7",
    "pydantic>=2.12.5",
    "pymongo",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
ingest = "main:main"

[tool.pylint.messages_control]
disable = [
    "C0114", # missing-module-docstring
    "C0116", # missing-function-docstring
    "E1111", # assignment-from-no-return (stubs)
    "W0613", # unused-argument (stubs)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = [".."]
